+++ springcloudcmp/ssh-init.sh
@@ -17,16 +17,20 @@ for i in "$@"
 do
  echo =======$i=======
  ssh-copy-id -i ~/.ssh/id_rsa.pub $i
- expect <<-EOF
- set timeout -1
- spawn  ssh-copy-id -i /root/.ssh/id_rsa.pub $i
- expect {
-  "*yes/no" { send "yes\n"; exp_continue }
-  "*exist" { send "login ok\n" }
-  "*password" { send "${passwd}\n" }
- }
-expect eof
-EOF
+ if [ $? -eq 1 ]; then
+        echo "你已多次输错密码，脚本将终止执行！"
+        exit 1
+ fi
+# expect <<-EOF
+# set timeout -1
+# spawn  ssh-copy-id -i /root/.ssh/id_rsa.pub $i
+# expect {
+#  "*yes/no" { send "yes\n"; exp_continue }
+# "*exist" { send "login ok\n" }
+#  "*password" { send "${passwd}\n" }
+# }
+#expect eof
+#EOF
 done
 
 rm -rf $wd
